// src/shared/context/context.d.ts
import { type ReactNode } from 'react';
import type { iProduto } from '../type/iProduto';
import type { iEstoque } from '../../features/estoque/type/iEstoque';
import type { iProdutoMov } from '../../shared/type/iProdutoMov';
import type { iFornecedor } from '../../shared/type/iFornecedor';
import type { iPedido } from '../../features/pedidos/type/iPedido';
interface IContext {
    setConexaoInternet: (internet: true | false) => void;
    conexaoInternet: true | false;
    setServidorOnline: (online: true | false) => void;
    servidorOnline: true | false;
    sessaoAtiva: true | false;
    setSessaoAtiva: (ativa: true | false) => void;
    listaTipoProdutos: iProduto[];
    setlistaTipoProdutos: (produto: iProduto[]) => void;
    categoria: string;
    setCategoria: (categoria: string) => void;
    listaProdutoEstoque: iProduto[];
    setListaProdutoEstoque: (item: iProduto[]) => void;
    contQtdEstoque: number;
    setContQtdEstoque: (cont: number) => void;
    contSuficiente: number;
    setContSuficiente: (cont: number) => void;
    contInsuficiente: number;
    setContInsuficiente: (cont: number) => void;
    estoqueSalvo: iEstoque | null;
    setEstoqueSalvo: (estoque: iEstoque | null) => void;
    listaProdutoMovTemp: iProdutoMov[];
    setListaProdutoMovTemp: (item: iProdutoMov[]) => void;
    listaProdutoMov: iProdutoMov[];
    setListaProdutoMov: (item: iProdutoMov[]) => void;
    tipoModal: string;
    setTipoModal: (mov: string) => void;
    handleModal: boolean;
    setHandleModal: (handle: boolean) => void;
    listaHistoricoMovEstoque: iProdutoMov[];
    setListaHistoricoMovEstoque: (item: iProdutoMov[]) => void;
    listaTipoMovimentacoes: string[];
    setListaTipoMovimentacoes: (mov: string[]) => void;
    listaMovimentacoesEstoque: string[];
    setListaMovimentacoesEstoque: (mov: string[]) => void;
    tipoMovSelecionado: string;
    setTipoMovSelecionado: (mov: string) => void;
    movimentacaoSelecionada: string;
    setMovimentacaoSelecionada: (mov: string) => void;
    tipoEntrada: string | null;
    setTipoEntrada: (tipo: string | null) => void;
    tipoSaida: string | null;
    setTipoSaida: (tipo: string | null) => void;
    mostrarCaixaDialogo: boolean;
    setMostrarCaixaDialogo: (mostrar: boolean) => void;
    tipoInput: "auto" | "manual";
    setTipoInput: (tipo: "auto" | "manual") => void;
    estoqueId: number;
    setEstoqueId: (id: number) => void;
    listaFornecedores: iFornecedor[];
    setListaFornecedores: (fornecedores: iFornecedor[]) => void;
    listaPedidosCompra: iPedido[];
    setListaPedidosCompra: (pedidos: iPedido[]) => void;
    origemDoModal: 'paginaPedido' | 'modalCriarPedido';
    setOrigemDoModal: (origem: 'paginaPedido' | 'modalCriarPedido') => void;
}
interface AppProvideProps {
    children: ReactNode;
}
export declare const AppContext: import("react").Context<IContext>;
export declare const unidadesMedida: string[];
export declare const movimentacoesEstoque: string[];
export declare const AppProvider: ({ children }: AppProvideProps) => import("react/jsx-runtime").JSX.Element;
export {};